[feat] complete Stage 1
Basic skeleton of the game that takes two numbers as input - an int for the
size of the board and a long for the seed for Random().

Write the class for the board with create() and print() methods to populate
the game board with boolean values and output either 'O' (letter o) or
' ' to the console.